+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - rename ActiveRecordCompose::InnerModel to ActiveRecordCompose::WrappedModel
+- rename ActiveRecordCompose::InnerModelCollection to ActiveRecordCompose::ComposedCollection
 
 ## [0.6.3] - 2025-01-31